Struct ivm::inet::INet

pub struct INet {
    pub nodes: Vec<Node>,
    free_nodes: Vec<NodeIdx>,
}

This represents the interaction net

Fields
nodes: Vec<Node>free_nodes: Vec<NodeIdx>
Implementations

impl INet

pub fn new_node(&mut self, agent_id: AgentId, port_count: usize) -> NodeIdx

Allocate a new node

pub fn free_node(&mut self, node_idx: NodeIdx)

Make node available for reuse

fn link(&mut self, a: NodePort, b: NodePort)

Mutually link ports

pub fn validate(&self, allow_unused_nodes: bool)

Validate the inet, panics if invalid, useful for debugging/tests If an INet generated from a valid AST fails validation, it’d be a bug

fn add_connector<'a, 'b: 'a>( &mut self, connector: &'b Connector, deferred_port_links: &'a mut Vec<[Result<NodePort, PortNameRef<'b>>; 2]>, agent_name_to_id: &HashMap<AgentName, AgentId>, created_nodes: &mut NodeIndices ) -> Result<NodePort, PortNameRef<'b>>

Add a Connector to the net, only called by INet::add_connections. If the connector is a port, Err(port_name) is returned so that it can be linked later. If the connector is an agent, it is added as a new node and its principal port is returned as Ok(port). Connections to auxiliary ports are added to deferred_port_links so that they can be linked later. This function is only called by INet::add_connections

pub fn add_connections<'a>( &mut self, connections: &'a [Connection], external_ports: Vec<[Result<NodePort, PortNameRef<'a>>; 2]>, agent_name_to_id: &HashMap<AgentName, AgentId> ) -> NodeIndices

Add connections to the net, either when creating a net from scratch or inserting a sub-net during a rewrite. external_ports is a map from port names to ports in the parent net. When creating a net from scratch (in ValidatedAst::into_inet_program), external_ports only contains the root port. When called by RuleBook::apply_rewrite_rule, external_ports contains the ports of the parent net that the rule’s RHS sub-net is connected to, based on which INet ports the rule’s LHS (active pair) aux port names map to, in that context of the local rewrite. Returns the indices of the created nodes.

pub fn remove_unused_nodes(&mut self)

Remove all unused nodes, called by RuleBook::pre_preduce

const INTERMEDIARY_AGENT_ID: AgentId = 4_294_967_295u32

fn rewrite_active_pair( &mut self, (a, b): ActiveNodePair, rule_book: &RuleBook, reuse: &mut ReuseableRewriteData ) -> bool

Rewrite active pair using rule book. Assumes that reuse buffers are empty. Returns false if active pair could not be rewritten because there was no matching rule Returns true when a rule matched and a rewrite happened, in that case reuse.inet_new_active_pairs_created_by_rewrite contains the new active pairs.

pub fn insert_rule_rhs_subnet( &mut self, subnet: &INet, reuse: &mut ReuseableSubnetData )

Inserts a rule’s RHS sub-net into the net, called by RuleBook::apply_rewrite_rule. It works by:

Creating nodes in self for the subnet nodes, to get new node indices

Copying ports of subnet nodes to the new nodes, translating links using new node indices

Pass-through-linking the subnet root node’s ports to the corresponding ports in self

Removing the subnet root node from self The aux ports of the two active pair nodes are mapped to the subnet’s root node ports: external_ports maps subnet root node port indices to ports in the self net. External (subnet root node) port indices refer to auxiliary ports in the active pair of the rule LHS: E.g. if the rule LHS is X(a, b) ~ Y(c, d), then external_ports contains the ports of self that map to [a, b, c, d] (in that order) in the context of the local rewrite. In other words, the ports are indexed left-to-right, after ordering both active pair agent IDs: So the external ports would be the same if the rule LHS was written as Y(c, d) ~ X(a, b). (Assuming X’s AgentId < Y’s AgentId in this example.) This function assumes that reuse.subnet_node_idx_to_main_node_idx is empty and that reuse.rule_book_external_ports contains the external ports of the subnet. When this function returns, reuse.inet_subnet_node_idx_to_main_node_idx contains the node indices of the inserted subnet nodes, to enable rewrite_active_pair to determine new active pairs. (reuse.inet_subnet_node_idx_to_main_node_idx[0] will be the index of the removed subnet root node.)

fn node_is_part_of_active_pair(&self, node_idx: NodeIdx) -> Option<NodeIdx>

Determines if a given node is part of an active pair and returns the other node in the pair

pub fn scan_active_pairs(&self) -> ActiveNodePairs

Determine active pairs that can potentially be rewritten if there is a matching rule

fn nodes_whose_principal_port_points_outside_subnet(&self) -> NodeIndices

Returns the active pairs that can come into existence after inserting a subnet, pass-through-linking through its root node and removing it. In other words, returns indices of nodes whose principal port points to a root node port.

pub fn active_pair_candidates_after_inserting_subnet( &self, rule_book: &RuleBook ) -> ActivePairCandidates

Returns the active pairs that can come into existence after inserting a subnet:

Field active_pairs_inside_subnet contains the active pairs that already exist in the subnet, which will definitely exist after inserting the subnet (and for which a rule exists)

Field nodes_whose_principal_port_points_outside_subnet contains the indices of nodes whose principal port points to a root node port, which can form active pairs with other nodes whose principal port points to aux ports of an active pair that gets rewritten. So these are not definitive active pair nodes, but they could form active pairs after a rewrite.

pub fn scan_active_pairs_and_reduce_step( &mut self, rule_book: &RuleBook ) -> bool

Perform one reduction step: Scan for active pairs and rewrite the first one for which a rule exists. Returns true if a rewrite happened, false otherwise (either because there are no active pairs or because no rule applies to any of them). Not efficient because of the rescan for active pairs, but useful for debugging.

pub fn reduce(&mut self, rule_book: &RuleBook) -> usize

Reduce net until no more reductions are possible, without rescanning for active pairs after each rewrite. Only scans the net for active pairs in the beginning. After each rewrite, new active pairs are found by checking the nodes involved in and adjacent to the rewritten sub-net.

pub fn reduce_in_max_steps<const MAX_REDUCTIONS: usize>( &mut self, rule_book: &RuleBook ) -> Option<usize>

Like reduce but performs maximum max_reductions reduction steps. Returns the number of reductions actually performed, None if limit was reached.

pub fn read_back_raw( &self, agent_id_to_name: &HashMap<AgentId, AgentName>, root_port_names: &[PortName] ) -> Vec<Connection>

Read back reduced net into textual form raw (unchanged)

pub fn read_back( &self, agent_id_to_name: &HashMap<AgentId, AgentName>, root_port_names: &[PortName] ) -> String

Read back reduced net into readable (nested) textual form