Module ivm::parser::ast

Structs

ActivePair

E.g. A(a, b) ~ B(c, d)

Agent

E.g. A(b, c)

Ast

AST of source file

Connection

E.g. A(a, b) ~ c

NestedAgent

E.g. A(a, B(b, c))

NestedConnection

E.g. A(a, B(b, c)) ~ C(D(d, e), f)

Rule

A rule has a LHS and RHS, and the LHS of a rule is an active pair that has its own LHS and RHS E.g. rule Add(ret, a) ~ Succ(b) = ret ~ Succ(cnt), Add(cnt, a) ~ b

Spanned

Wrapper for a value with a span

ValidatedAst

Enums

Connector

One end of a connection, can be either agent or port

NestedConnector

Like Connector, but can contain nested agents E.g. A(a, B(b, c))

Constants

INIT_CONNECTIONS

Matches Token::KeywordInit

ROOT_NODE_IDX

The first node in a net is the root node

ROOT_PORT_NAME

Functions

process_agents_and_ports

Iterate over connections and process agents and ports with the given functions

validate_agent_port_references 🔒

Reject duplicate port names in agents of active pair E.g. A(a, a) ~ B is invalid, port names in active pair must be distinct Returns set of port names in agent

validate_agent_usage 🔒

Validate agent usage, returns false if agent is undeclared

validate_connections 🔒

Validate connections, either in a rule’s RHS or in the initial net: Check linearity restriction (each port must be referenced exactly twice) Check that each agent usage has correct number of ports matching its declaration rule_src is used for error messages, pass None when validating init connections

Type Definitions

AgentName

Starts with uppercase letter (Token::Agent)

PortName

Starts with lowercase letter (Token::Port)

PortNameRef

Opaque Types

🔒