Module ivm::parser::ast

Structs

ActivePair

E.g. A(a, b) ~ B(c, d)

Agent

E.g. A(b, c)

Ast

AST of source file

Connection

E.g. A(a, b) ~ c

NestedAgent

E.g. A(a, B(b, c))

NestedConnection

E.g. A(a, B(b, c)) ~ C(D(d, e), f)

Rule

A rule has a LHS and RHS, and the LHS of a rule is an active pair that has its own LHS and RHS E.g. rule Add(ret, a) ~ Succ(b) = ret ~ Succ(cnt), Add(cnt, a) ~ b

Spanned

Wrapper for a value with a span

ValidatedAst

A validated AST and its corresponding RuleBook

Enums

Connector

One end of a connection, can be either agent or port

NestedConnector

Like Connector, but can contain nested agents E.g. A(a, B(b, c))

Constants

INIT_CONNECTIONS

Matches Token::KeywordInit

ROOT_NODE_IDX

The first node in a net is the root node

ROOT_PORT_NAME

Port name of the root node in init connections

Functions

process_agents_and_ports

Iterate over connections and process agents and ports with the given functions

Type Definitions

AgentName

Starts with uppercase letter (Token::Agent)

PortName

Starts with lowercase letter (Token::Port)

PortNameRef